nlp/tools/gemini-api/requirements.txt
streamlit
google-generativeai
pymongo
voyageai
python-dotenv
langchain-text-splitters